% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Galton.R
\docType{data}
\name{Galton2}
\alias{Galton2}
\title{The famous Galton data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 898 rows and 6 columns.}
\usage{
Galton2
}
\description{
These data were collected by Galton, inventor of the correlation coefficient.
}
\examples{

head(Galton)

}
\keyword{datasets}
