% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-package.R
\name{start_a_package}
\alias{start_a_package}
\title{Favoraite package startup stuff}
\usage{
start_a_package(use_build_ignore = FALSE)
}
\arguments{
\item{use_build_ignore}{A logical indicating whether I want to use a build ingore.}
}
\value{
Nothing.  Use for side-effects
}
\description{
Run a bunch of things from `devtools` to stub out a package.
}
\examples{
start_a_package()

}
